backend/requirements.txt
fastapi
httpx
redis
python-dotenv
pydantic
